Add transactable trait
