+++ automerge-wasm/src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::unused_unit)]
 use am::transaction::CommitOptions;
+use am::transaction::Transactable;
 use automerge as am;
 use automerge::{Change, ObjId, Prop, Value, ROOT};
 use js_sys::{Array, Object, Uint8Array};
