+++ automerge/tests/test.rs
@@ -1,3 +1,4 @@
+use automerge::transaction::Transactable;
 use automerge::{ActorId, AutoTxn, Automerge, Value, ROOT};
 
 mod helpers;
